move `ComponentData` to its own module so it can be reused
