+++ cyclonedx-bom/src/specs/v1_5/modelcard.rs
@@ -25,9 +25,8 @@ use xml::{
 
 use crate::{
     errors::XmlReadError,
-    models::{self, bom::BomReference},
-    prelude::Uri,
-    specs::{common::property::Properties, v1_5::data_governance::DataGovernance},
+    models,
+    specs::common::property::Properties,
     utilities::{convert_optional, convert_vec},
     xml::{
         optional_attribute, read_simple_tag, to_xml_read_error, to_xml_write_error,
@@ -36,7 +35,7 @@ use crate::{
     },
 };
 
-use super::attachment::Attachment;
+use super::component_data::{ComponentData, GraphicsCollection};
 
 #[derive(Debug, Deserialize, Serialize, PartialEq)]
 #[serde(rename_all = "camelCase")]
@@ -68,7 +67,7 @@ impl From<models::modelcard::ModelCard> for ModelCard {
 impl From<ModelCard> for models::modelcard::ModelCard {
     fn from(other: ModelCard) -> Self {
         Self {
-            bom_ref: other.bom_ref.map(BomReference::new),
+            bom_ref: other.bom_ref.map(models::bom::BomReference::new),
             model_parameters: convert_optional(other.model_parameters),
             quantitative_analysis: convert_optional(other.quantitative_analysis),
             considerations: convert_optional(other.considerations),
@@ -230,7 +229,6 @@ const INPUT_TAG: &str = "input";
 const OUTPUTS_TAG: &str = "outputs";
 const OUTPUT_TAG: &str = "output";
 const FORMAT_TAG: &str = "format";
-const ATTACHMENT_TAG: &str = "attachment";
 
 impl ToXml for ModelParameters {
     fn write_xml_element<W: std::io::Write>(
@@ -541,12 +539,7 @@ impl From<Dataset> for models::modelcard::Dataset {
 
 const DATASETS_TAG: &str = "datasets";
 const DATASET_TAG: &str = "dataset";
-const CONTENTS_TAG: &str = "contents";
 const GRAPHICS_TAG: &str = "graphics";
-const NAME_TAG: &str = "name";
-const CLASSIFICATION_TAG: &str = "classification";
-const SENSITIVE_DATA_TAG: &str = "sensitiveData";
-const GOVERNANCE_TAG: &str = "governance";
 const REF_TAG: &str = "ref";
 
 impl ToXml for Dataset {
@@ -556,7 +549,7 @@ impl ToXml for Dataset {
     ) -> Result<(), crate::errors::XmlWriteError> {
         match self {
             Dataset::Component(component) => {
-                component.write_xml_element(writer)?;
+                component.write_xml_named_element(writer, DATASET_TAG)?;
             }
             Dataset::Reference(reference) => {
                 write_start_tag(writer, DATASET_TAG)?;
@@ -578,327 +571,13 @@ impl FromXml for Dataset {
     where
         Self: Sized,
     {
-        let bom_ref = optional_attribute(attributes, BOM_REF_ATTR);
-        let mut data_type: Option<String> = None;
-        let mut data_name: Option<String> = None;
-        let mut contents: Option<DataContents> = None;
-        let mut classification: Option<String> = None;
-        let mut graphics: Option<Graphics> = None;
-        let mut description: Option<String> = None;
-        let mut governance: Option<DataGovernance> = None;
-        let mut sensitive_data: Option<String> = None;
-
-        let mut got_end_tag = false;
-        while !got_end_tag {
-            let next_element = event_reader
-                .next()
-                .map_err(to_xml_read_error(DATASET_TAG))?;
-
-            match next_element {
-                reader::XmlEvent::StartElement { name, .. } if name.local_name == TYPE_TAG => {
-                    data_type = Some(read_simple_tag(event_reader, &name)?);
-                }
-
-                reader::XmlEvent::StartElement { name, .. } if name.local_name == NAME_TAG => {
-                    data_name = Some(read_simple_tag(event_reader, &name)?);
-                }
-
-                reader::XmlEvent::StartElement {
-                    name, attributes, ..
-                } if name.local_name == CONTENTS_TAG => {
-                    contents = Some(DataContents::read_xml_element(
-                        event_reader,
-                        &name,
-                        &attributes,
-                    )?);
-                }
-
-                reader::XmlEvent::StartElement { name, .. }
-                    if name.local_name == DESCRIPTION_TAG =>
-                {
-                    description = Some(read_simple_tag(event_reader, &name)?);
-                }
-
-                reader::XmlEvent::StartElement { name, .. }
-                    if name.local_name == CLASSIFICATION_TAG =>
-                {
-                    classification = Some(read_simple_tag(event_reader, &name)?);
-                }
-
-                reader::XmlEvent::StartElement {
-                    name, attributes, ..
-                } if name.local_name == GOVERNANCE_TAG => {
-                    governance = Some(DataGovernance::read_xml_element(
-                        event_reader,
-                        &name,
-                        &attributes,
-                    )?);
-                }
-
-                reader::XmlEvent::StartElement {
-                    name, attributes, ..
-                } if name.local_name == GRAPHICS_TAG => {
-                    graphics = Some(Graphics::read_xml_element(
-                        event_reader,
-                        &name,
-                        &attributes,
-                    )?);
-                }
-
-                reader::XmlEvent::StartElement { name, .. }
-                    if name.local_name == SENSITIVE_DATA_TAG =>
-                {
-                    sensitive_data = Some(read_simple_tag(event_reader, &name)?);
-                }
-
-                reader::XmlEvent::EndElement { name } if &name == element_name => {
-                    got_end_tag = true;
-                }
-
-                _ => (),
-            }
-        }
-
-        let data_type = data_type.ok_or_else(|| XmlReadError::RequiredDataMissing {
-            required_field: TYPE_TAG.to_string(),
-            element: element_name.local_name.to_string(),
-        })?;
-
-        Ok(Self::Component(ComponentData {
-            bom_ref,
-            data_type,
-            name: data_name,
-            contents,
-            classification,
-            sensitive_data,
-            graphics,
-            description,
-            governance,
-        }))
-    }
-}
-
-/// Dataset component, for more details see:
-/// https://cyclonedx.org/docs/1.5/json/#tab-pane_components_items_modelCard_modelParameters_datasets_items_oneOf_i1
-#[derive(Debug, Deserialize, Serialize, PartialEq)]
-#[serde(rename_all = "camelCase")]
-pub(crate) struct ComponentData {
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub(crate) bom_ref: Option<String>,
-    #[serde(rename = "type")]
-    pub(crate) data_type: String,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub(crate) name: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub(crate) contents: Option<DataContents>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub(crate) classification: Option<String>,
-    /// Marked as an array of `String`, but examples use a single entry
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub(crate) sensitive_data: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub(crate) graphics: Option<Graphics>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub(crate) description: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub(crate) governance: Option<DataGovernance>,
-}
-
-impl From<models::modelcard::ComponentData> for ComponentData {
-    fn from(other: models::modelcard::ComponentData) -> Self {
-        Self {
-            bom_ref: other.bom_ref.map(|r| r.0),
-            data_type: other.data_type.to_string(),
-            name: other.name,
-            contents: convert_optional(other.contents),
-            classification: convert_optional(other.classification),
-            sensitive_data: convert_optional(other.sensitive_data),
-            graphics: convert_optional(other.graphics),
-            description: convert_optional(other.description),
-            governance: convert_optional(other.governance),
-        }
-    }
-}
-
-impl From<ComponentData> for models::modelcard::ComponentData {
-    fn from(other: ComponentData) -> Self {
-        Self {
-            bom_ref: other.bom_ref.map(BomReference::new),
-            data_type: models::modelcard::ComponentDataType::new_unchecked(other.data_type),
-            name: other.name,
-            contents: convert_optional(other.contents),
-            classification: convert_optional(other.classification),
-            sensitive_data: convert_optional(other.sensitive_data),
-            graphics: convert_optional(other.graphics),
-            description: convert_optional(other.description),
-            governance: convert_optional(other.governance),
-        }
-    }
-}
-
-impl ToXml for ComponentData {
-    fn write_xml_element<W: std::io::Write>(
-        &self,
-        writer: &mut xml::EventWriter<W>,
-    ) -> Result<(), crate::errors::XmlWriteError> {
-        let mut start_tag = writer::XmlEvent::start_element(DATASET_TAG);
-        if let Some(bom_ref) = &self.bom_ref {
-            start_tag = start_tag.attr(BOM_REF_ATTR, bom_ref);
-        }
-        writer
-            .write(start_tag)
-            .map_err(to_xml_write_error(DATASET_TAG))?;
-
-        if let Some(name) = &self.name {
-            write_simple_tag(writer, NAME_TAG, name)?;
-        }
-
-        if let Some(contents) = &self.contents {
-            contents.write_xml_element(writer)?;
-        }
-
-        if let Some(classification) = &self.classification {
-            write_simple_tag(writer, CLASSIFICATION_TAG, classification)?;
-        }
-
-        if let Some(sensitive_data) = &self.sensitive_data {
-            write_simple_tag(writer, SENSITIVE_DATA_TAG, sensitive_data)?;
-        }
-
-        if let Some(graphics) = &self.graphics {
-            graphics.write_xml_element(writer)?;
-        }
-
-        if let Some(description) = &self.description {
-            write_simple_tag(writer, DESCRIPTION_TAG, description)?;
-        }
-
-        if let Some(governance) = &self.governance {
-            governance.write_xml_named_element(writer, GOVERNANCE_TAG)?;
-        }
-
-        write_close_tag(writer, DATASET_TAG)?;
-
-        Ok(())
-    }
-}
-
-#[derive(Debug, Deserialize, Serialize, PartialEq)]
-#[serde(rename_all = "camelCase")]
-pub(crate) struct DataContents {
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub(crate) attachment: Option<Attachment>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub(crate) url: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub(crate) properties: Option<Properties>,
-}
-
-impl From<models::modelcard::DataContents> for DataContents {
-    fn from(other: models::modelcard::DataContents) -> Self {
-        Self {
-            attachment: convert_optional(other.attachment),
-            url: other.url.map(|url| url.to_string()),
-            properties: convert_optional(other.properties),
-        }
-    }
-}
-
-impl From<DataContents> for models::modelcard::DataContents {
-    fn from(other: DataContents) -> Self {
-        Self {
-            attachment: convert_optional(other.attachment),
-            url: other.url.map(Uri),
-            properties: convert_optional(other.properties),
-        }
-    }
-}
-
-impl ToXml for DataContents {
-    fn write_xml_element<W: std::io::Write>(
-        &self,
-        writer: &mut xml::EventWriter<W>,
-    ) -> Result<(), crate::errors::XmlWriteError> {
-        write_start_tag(writer, CONTENTS_TAG)?;
-
-        if let Some(attachment) = &self.attachment {
-            attachment.write_xml_named_element(writer, ATTACHMENT_TAG)?;
-        }
-
-        if let Some(url) = &self.url {
-            write_simple_tag(writer, URL_TAG, url)?;
-        }
-
-        if let Some(properties) = &self.properties {
-            properties.write_xml_element(writer)?;
-        }
-
-        write_close_tag(writer, CONTENTS_TAG)?;
-
-        Ok(())
-    }
-}
-
-const URL_TAG: &str = "url";
-const PROPERTIES_TAG: &str = "properties";
-
-impl FromXml for DataContents {
-    fn read_xml_element<R: std::io::Read>(
-        event_reader: &mut xml::EventReader<R>,
-        element_name: &OwnedName,
-        _attributes: &[xml::attribute::OwnedAttribute],
-    ) -> Result<Self, XmlReadError>
-    where
-        Self: Sized,
-    {
-        let mut url: Option<String> = None;
-        let mut attachment: Option<Attachment> = None;
-        let mut properties: Option<Properties> = None;
-
-        let mut got_end_tag = false;
-        while !got_end_tag {
-            let next_element = event_reader
-                .next()
-                .map_err(to_xml_read_error(&element_name.local_name))?;
-
-            match next_element {
-                reader::XmlEvent::StartElement { name, .. } if name.local_name == URL_TAG => {
-                    url = Some(read_simple_tag(event_reader, &name)?)
-                }
-
-                reader::XmlEvent::StartElement {
-                    name, attributes, ..
-                } if name.local_name == ATTACHMENT_TAG => {
-                    attachment = Some(Attachment::read_xml_element(
-                        event_reader,
-                        &name,
-                        &attributes,
-                    )?);
-                }
-
-                reader::XmlEvent::StartElement {
-                    name, attributes, ..
-                } if name.local_name == PROPERTIES_TAG => {
-                    properties = Some(Properties::read_xml_element(
-                        event_reader,
-                        &name,
-                        &attributes,
-                    )?)
-                }
-
-                reader::XmlEvent::EndElement { name } if &name == element_name => {
-                    got_end_tag = true;
-                }
-
-                _ => (),
-            }
-        }
-
-        Ok(Self {
-            attachment,
-            url,
-            properties,
-        })
+        // Reference is deprecated, so we don't attempt to read it (it's not even part of the
+        // json.schema specification)
+        Ok(Self::Component(ComponentData::read_xml_element(
+            event_reader,
+            element_name,
+            attributes,
+        )?))
     }
 }
 
@@ -906,7 +585,7 @@ impl FromXml for DataContents {
 #[serde(rename_all = "camelCase")]
 pub(crate) struct QuantitativeAnalysis {
     pub(crate) performance_metrics: Option<PerformanceMetrics>,
-    pub(crate) graphics: Option<Graphics>,
+    pub(crate) graphics: Option<GraphicsCollection>,
 }
 
 impl From<models::modelcard::QuantitativeAnalysis> for QuantitativeAnalysis {
@@ -943,7 +622,7 @@ impl ToXml for QuantitativeAnalysis {
         }
 
         if let Some(graphics) = &self.graphics {
-            graphics.write_xml_element(writer)?;
+            graphics.write_xml_named_element(writer, GRAPHICS_TAG)?;
         }
 
         write_close_tag(writer, QUANTITATIVE_ANALYSIS_TAG)?;
@@ -962,7 +641,7 @@ impl FromXml for QuantitativeAnalysis {
         Self: Sized,
     {
         let mut performance_metrics: Option<PerformanceMetrics> = None;
-        let mut graphics: Option<Graphics> = None;
+        let mut graphics: Option<GraphicsCollection> = None;
 
         let mut got_end_tag = false;
         while !got_end_tag {
@@ -984,7 +663,7 @@ impl FromXml for QuantitativeAnalysis {
                 reader::XmlEvent::StartElement {
                     name, attributes, ..
                 } if name.local_name == GRAPHICS_TAG => {
-                    graphics = Some(Graphics::read_xml_element(
+                    graphics = Some(GraphicsCollection::read_xml_element(
                         event_reader,
                         &name,
                         &attributes,
@@ -1555,285 +1234,25 @@ impl FromXml for MLParameter {
     }
 }
 
-/// For more details see:
-/// https://cyclonedx.org/docs/1.5/json/#components_items_modelCard_modelParameters_datasets_items_oneOf_i0_graphics
-#[derive(Debug, Deserialize, Serialize, PartialEq)]
-pub(crate) struct Graphics {
-    pub(crate) description: Option<String>,
-    pub(crate) collection: Option<Collection>,
-}
-
-impl From<models::modelcard::GraphicsCollection> for Graphics {
-    fn from(other: models::modelcard::GraphicsCollection) -> Self {
-        Self {
-            description: convert_optional(other.description),
-            collection: convert_optional(other.collection),
-        }
-    }
-}
-
-impl From<Graphics> for models::modelcard::GraphicsCollection {
-    fn from(other: Graphics) -> Self {
-        Self {
-            description: convert_optional(other.description),
-            collection: convert_optional(other.collection),
-        }
-    }
-}
-
-const COLLECTION_TAG: &str = "collection";
-const DESCRIPTION_TAG: &str = "description";
-
-impl ToXml for Graphics {
-    fn write_xml_element<W: std::io::Write>(
-        &self,
-        writer: &mut xml::EventWriter<W>,
-    ) -> Result<(), crate::errors::XmlWriteError> {
-        write_start_tag(writer, GRAPHICS_TAG)?;
-
-        if let Some(description) = &self.description {
-            write_simple_tag(writer, DESCRIPTION_TAG, description)?;
-        }
-
-        if let Some(collection) = &self.collection {
-            collection.write_xml_element(writer)?;
-        }
-
-        write_close_tag(writer, GRAPHICS_TAG)?;
-
-        Ok(())
-    }
-}
-
-impl FromXml for Graphics {
-    fn read_xml_element<R: std::io::Read>(
-        event_reader: &mut xml::EventReader<R>,
-        element_name: &OwnedName,
-        _attributes: &[xml::attribute::OwnedAttribute],
-    ) -> Result<Self, XmlReadError>
-    where
-        Self: Sized,
-    {
-        let mut description: Option<String> = None;
-        let mut collection: Option<Collection> = None;
-
-        let mut got_end_tag = false;
-        while !got_end_tag {
-            let next_element = event_reader.next().map_err(to_xml_read_error(OUTPUT_TAG))?;
-            match next_element {
-                reader::XmlEvent::StartElement { name, .. }
-                    if name.local_name == DESCRIPTION_TAG =>
-                {
-                    description = Some(read_simple_tag(event_reader, &name)?);
-                }
-
-                reader::XmlEvent::StartElement {
-                    name, attributes, ..
-                } if name.local_name == COLLECTION_TAG => {
-                    collection = Some(Collection::read_xml_element(
-                        event_reader,
-                        &name,
-                        &attributes,
-                    )?);
-                }
-
-                reader::XmlEvent::EndElement { name } if &name == element_name => {
-                    got_end_tag = true;
-                }
-
-                _ => (),
-            }
-        }
-
-        Ok(Self {
-            description,
-            collection,
-        })
-    }
-}
-
-/// Helper struct to collect all [`Graphic`].
-#[derive(Debug, Deserialize, Serialize, PartialEq)]
-pub(crate) struct Collection(pub(crate) Vec<Graphic>);
-
-impl From<Vec<models::modelcard::Graphic>> for Collection {
-    fn from(other: Vec<models::modelcard::Graphic>) -> Self {
-        Self(convert_vec(other))
-    }
-}
-
-impl From<Collection> for Vec<models::modelcard::Graphic> {
-    fn from(other: Collection) -> Self {
-        convert_vec(other.0)
-    }
-}
-
-const GRAPHIC_TAG: &str = "graphic";
-
-impl ToXml for Collection {
-    fn write_xml_element<W: std::io::Write>(
-        &self,
-        writer: &mut xml::EventWriter<W>,
-    ) -> Result<(), crate::errors::XmlWriteError> {
-        write_start_tag(writer, COLLECTION_TAG)?;
-
-        for graphic in &self.0 {
-            graphic.write_xml_element(writer)?;
-        }
-
-        write_close_tag(writer, COLLECTION_TAG)?;
-
-        Ok(())
-    }
-}
-
-impl FromXml for Collection {
-    fn read_xml_element<R: std::io::Read>(
-        event_reader: &mut xml::EventReader<R>,
-        element_name: &OwnedName,
-        _attributes: &[xml::attribute::OwnedAttribute],
-    ) -> Result<Self, XmlReadError>
-    where
-        Self: Sized,
-    {
-        let mut collection: Vec<Graphic> = Vec::new();
-        let mut got_end_tag = false;
-
-        while !got_end_tag {
-            let next_element = event_reader.next().map_err(to_xml_read_error(OUTPUT_TAG))?;
-
-            match next_element {
-                reader::XmlEvent::StartElement {
-                    name, attributes, ..
-                } if name.local_name == GRAPHIC_TAG => {
-                    collection.push(Graphic::read_xml_element(event_reader, &name, &attributes)?);
-                }
-
-                reader::XmlEvent::EndElement { name } if &name == element_name => {
-                    got_end_tag = true;
-                }
-
-                _ => (),
-            }
-        }
-
-        Ok(Self(collection))
-    }
-}
-
-#[derive(Debug, Deserialize, Serialize, PartialEq)]
-pub(crate) struct Graphic {
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub(crate) name: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub(crate) image: Option<Attachment>,
-}
-
-impl From<models::modelcard::Graphic> for Graphic {
-    fn from(other: models::modelcard::Graphic) -> Self {
-        Self {
-            name: convert_optional(other.name),
-            image: convert_optional(other.image),
-        }
-    }
-}
-
-impl From<Graphic> for models::modelcard::Graphic {
-    fn from(other: Graphic) -> Self {
-        Self {
-            name: convert_optional(other.name),
-            image: convert_optional(other.image),
-        }
-    }
-}
-
-const IMAGE_TAG: &str = "image";
-
-impl ToXml for Graphic {
-    fn write_xml_element<W: std::io::Write>(
-        &self,
-        writer: &mut xml::EventWriter<W>,
-    ) -> Result<(), crate::errors::XmlWriteError> {
-        write_start_tag(writer, GRAPHIC_TAG)?;
-
-        if let Some(name) = &self.name {
-            write_simple_tag(writer, NAME_TAG, name)?;
-        }
-
-        if let Some(image) = &self.image {
-            image.write_xml_named_element(writer, IMAGE_TAG)?;
-        }
-
-        write_close_tag(writer, GRAPHIC_TAG)?;
-
-        Ok(())
-    }
-}
-
-impl FromXml for Graphic {
-    fn read_xml_element<R: std::io::Read>(
-        event_reader: &mut xml::EventReader<R>,
-        element_name: &OwnedName,
-        _attributes: &[xml::attribute::OwnedAttribute],
-    ) -> Result<Self, XmlReadError>
-    where
-        Self: Sized,
-    {
-        let mut graphic_name: Option<String> = None;
-        let mut image: Option<Attachment> = None;
-
-        let mut got_end_tag = false;
-
-        while !got_end_tag {
-            let next_element = event_reader.next().map_err(to_xml_read_error(OUTPUT_TAG))?;
-            match next_element {
-                reader::XmlEvent::StartElement { name, .. } if name.local_name == NAME_TAG => {
-                    graphic_name = Some(read_simple_tag(event_reader, &name)?);
-                }
-
-                reader::XmlEvent::StartElement {
-                    name, attributes, ..
-                } if name.local_name == IMAGE_TAG => {
-                    image = Some(Attachment::read_xml_element(
-                        event_reader,
-                        &name,
-                        &attributes,
-                    )?);
-                }
-
-                reader::XmlEvent::EndElement { name } if &name == element_name => {
-                    got_end_tag = true;
-                }
-
-                _ => (),
-            }
-        }
-
-        Ok(Self {
-            name: graphic_name,
-            image,
-        })
-    }
-}
-
 #[cfg(test)]
 pub(crate) mod test {
     use pretty_assertions::assert_eq;
 
     use crate::{
-        models::{self, bom::BomReference},
+        models,
         prelude::{NormalizedString, Uri},
         specs::{
+            common::bom_reference::BomReference,
             common::organization::{OrganizationalContact, OrganizationalEntity},
-            v1_5::modelcard::{
-                Attachment, Collection, ComponentData, ConfidenceInterval, DataContents, Dataset,
-                Datasets, Graphic, Graphics, Inputs, MLParameter, ModelCard, ModelParameters,
-                ModelParametersApproach, Outputs, PerformanceMetric, PerformanceMetrics,
-                QuantitativeAnalysis,
-            },
             v1_5::{
+                attachment::Attachment,
+                component_data::{ComponentData, DataContents, Graphic, GraphicsCollection},
                 data_governance::{DataGovernance, DataGovernanceResponsibleParty},
-                modelcard::GOVERNANCE_TAG,
+                modelcard::{
+                    ConfidenceInterval, Dataset, Datasets, Inputs, MLParameter, ModelCard,
+                    ModelParameters, ModelParametersApproach, Outputs, PerformanceMetric,
+                    PerformanceMetrics, QuantitativeAnalysis,
+                },
             },
         },
         xml::test::{
@@ -1855,16 +1274,19 @@ pub(crate) mod test {
                         upper_bound: Some("high".to_string()),
                     }),
                 }])),
-                graphics: Some(Graphics {
+                graphics: Some(GraphicsCollection {
                     description: Some("Graphic Desc".to_string()),
-                    collection: Some(Collection(vec![Graphic {
-                        name: Some("Graphic A".to_string()),
-                        image: Some(Attachment {
-                            content: "1234".to_string(),
-                            content_type: None,
-                            encoding: None,
-                        }),
-                    }])),
+                    collection: Some(
+                        vec![Graphic {
+                            name: Some("Graphic A".to_string()),
+                            image: Some(Attachment {
+                                content: "1234".to_string(),
+                                content_type: None,
+                                encoding: None,
+                            }),
+                        }]
+                        .into(),
+                    ),
                 }),
             }),
             considerations: None,
@@ -1874,7 +1296,7 @@ pub(crate) mod test {
 
     pub(crate) fn corresponding_modelcard() -> models::modelcard::ModelCard {
         models::modelcard::ModelCard {
-            bom_ref: Some(BomReference::new("modelcard-1")),
+            bom_ref: Some(models::bom::BomReference::new("modelcard-1")),
             model_parameters: Some(corresponding_model_parameters()),
             quantitative_analysis: Some(models::modelcard::QuantitativeAnalysis {
                 performance_metrics: Some(models::modelcard::PerformanceMetrics(vec![
@@ -1927,7 +1349,7 @@ pub(crate) mod test {
             owners: Some(vec![
                 models::data_governance::DataGovernanceResponsibleParty::Contact(
                     models::organization::OrganizationalContact {
-                        bom_ref: Some(BomReference::new("contact-1")),
+                        bom_ref: Some(models::bom::BomReference::new("contact-1")),
                         name: Some(NormalizedString::new("Contact")),
                         email: Some(NormalizedString::new("contact@example.com")),
                         phone: None,
@@ -1946,7 +1368,7 @@ pub(crate) mod test {
             architecture_family: Some("Architecture".to_string()),
             model_architecture: Some("Model".to_string()),
             datasets: Some(Datasets(vec![Dataset::Component(ComponentData {
-                bom_ref: Some("dataset-1".to_string()),
+                bom_ref: Some(BomReference::new("dataset-1")),
                 data_type: "dataset".to_string(),
                 name: Some("Training Data".to_string()),
                 contents: Some(DataContents {
@@ -1975,7 +1397,7 @@ pub(crate) mod test {
             model_architecture: Some("Model".to_string()),
             datasets: Some(models::modelcard::Datasets(vec![
                 models::modelcard::Dataset::Component(models::modelcard::ComponentData {
-                    bom_ref: Some(BomReference::new("dataset-1")),
+                    bom_ref: Some(models::bom::BomReference::new("dataset-1")),
                     data_type: models::modelcard::ComponentDataType::Dataset,
                     name: Some("Training Data".to_string()),
                     contents: Some(models::modelcard::DataContents {
@@ -2007,6 +1429,7 @@ pub(crate) mod test {
 
     #[test]
     fn it_should_write_xml_data_governance() {
+        const GOVERNANCE_TAG: &str = "governance";
         let xml_output = write_named_element_to_string(example_governance(), GOVERNANCE_TAG);
         insta::assert_snapshot!(xml_output);
     }
@@ -2070,9 +1493,9 @@ pub(crate) mod test {
                     upper_bound: Some("The upper bound of the confidence interval".to_string())
                 })
             }])),
-            graphics: Some(Graphics {
+            graphics: Some(GraphicsCollection {
                 description: Some("Performance images".to_string()),
-                collection: Some(Collection(vec![Graphic {
+                collection: Some(vec![Graphic {
                     name: Some(
                         "FID vs CLIP Scores on 512x512 samples for different v1-versions"
                             .to_string(),
@@ -2082,7 +1505,7 @@ pub(crate) mod test {
                         content_type: Some("image/jpeg".to_string()),
                         encoding: Some("base64".to_string()),
                     }),
-                }])),
+                }].into()),
             }),
         };
         assert_eq!(expected, actual);
@@ -2137,19 +1560,23 @@ pub(crate) mod test {
   </collection>
 </graphics>
 "#;
-        let actual: Graphics = read_element_from_string(input);
-        let expected = Graphics {
+        let actual: GraphicsCollection = read_element_from_string(input);
+        let expected = GraphicsCollection {
             description: Some("Performance images".to_string()),
-            collection: Some(Collection(vec![Graphic {
-                name: Some(
-                    "FID vs CLIP Scores on 512x512 samples for different v1-versions".to_string(),
-                ),
-                image: Some(Attachment {
-                    content: "abcdefgh".to_string(),
-                    content_type: Some("image/jpeg".to_string()),
-                    encoding: Some("base64".to_string()),
-                }),
-            }])),
+            collection: Some(
+                vec![Graphic {
+                    name: Some(
+                        "FID vs CLIP Scores on 512x512 samples for different v1-versions"
+                            .to_string(),
+                    ),
+                    image: Some(Attachment {
+                        content: "abcdefgh".to_string(),
+                        content_type: Some("image/jpeg".to_string()),
+                        encoding: Some("base64".to_string()),
+                    }),
+                }]
+                .into(),
+            ),
         };
         assert_eq!(expected, actual);
     }
@@ -2246,7 +1673,7 @@ pub(crate) mod test {
 "#;
         let actual: Dataset = read_element_from_string(input);
         let expected = Dataset::Component(ComponentData {
-            bom_ref: Some("dataset-a".to_string()),
+            bom_ref: Some(BomReference::new("dataset-a")),
             data_type: "dataset".to_string(),
             name: Some("Training Data".to_string()),
             contents: Some(DataContents {
